[MP][Observability] Restructure subscribers into metrics/logging/telemetry subdirs
Reorganize subscribers for cleaner project layout:
  subscribers/metrics/l1.py, sm.py
  subscribers/logging/l1.py, sm.py
  subscribers/telemetry/  (future PR 3)

Update all imports in server, blend_server, blend_server_v2, and tests.
